Fix and re-enable WPT drag and drop tests
This CL fixes how the WPT drag and drop tests send the input actions to
the browser to make the events coherent with each other.

Currently, all the pointer events are being sent as individual actions
without maintaining context between calls. So, for example, the first
pointerDown event is being sent at the point (0, 0) instead of the
center of the dragElement. By chaining all the pointer events into one
single action, the user agent can execute the events as they were
intended.

Bug:1353435
Change-Id: I4dcb189dffc71c5cd601e3fb4147c49d72e393b7